Cleaned up files, and created util function for files
