+++ tests/e2e/verify-distribution/verifydistribution/verify_distribution_test.go
@@ -5,9 +5,9 @@ package verifydistribution
 
 import (
 	"bufio"
-	"errors"
 	"fmt"
 	"github.com/onsi/gomega"
+	. "github.com/verrazzano/verrazzano/pkg/files"
 	. "github.com/verrazzano/verrazzano/pkg/string"
 	"github.com/verrazzano/verrazzano/pkg/test/framework"
 	"github.com/verrazzano/verrazzano/tools/vz/pkg/helpers"
@@ -179,33 +179,6 @@ var _ = t.Describe("Verify VZ distribution", func() {
 		})
 	})
 })
-
-// GetMatchingFiles returns the filenames for files that match a regular expression.
-func GetMatchingFiles(rootDirectory string, fileMatchRe *regexp.Regexp) (fileMatches []string, err error) {
-	if len(rootDirectory) == 0 {
-		return nil, errors.New("GetMatchingFiles requires a rootDirectory")
-	}
-
-	if fileMatchRe == nil {
-		return nil, fmt.Errorf("GetMatchingFiles requires a regular expression")
-	}
-
-	walkFunc := func(fileName string, fileInfo os.FileInfo, err error) error {
-		if !fileMatchRe.MatchString(fileName) {
-			return nil
-		}
-		if !fileInfo.IsDir() {
-			fileMatches = append(fileMatches, fileName)
-		}
-		return nil
-	}
-
-	err = filepath.Walk(rootDirectory, walkFunc)
-	if err != nil {
-		return nil, err
-	}
-	return fileMatches, err
-}
 
 // verifyDistributionByDirectory verifies the contents of inputDir with Values from map
 func verifyDistributionByDirectory(inputDir string, key string, variant string) {
